refactor: builder dir
