+++ packages/waku/src/dev.ts
@@ -1,4 +1,4 @@
 export { honoMiddleware } from './lib/middleware/hono-dev.js';
 export { connectMiddleware } from './lib/middleware/connect-dev.js';
 export { createHandler as unstable_createHandler } from './lib/rsc/handler-dev.js';
-export { build } from './lib/builder.js';
+export { build } from './lib/builder/build.js';
